+++ pubnet/network/_edge/numpy_edge.py
@@ -102,6 +102,19 @@ class NumpyEdge(Edge):
         with open(header_name, "wt") as header_file:
             header_file.write(header)
 
+    def _renumber_column(self, col: str, id_map: dict[int, int]):
+        if self.start_id == self.end_id:
+            self._data = np.fromiter(
+                ((id_map[row[0]], id_map[row[1]]) for row in self._data),
+                dtype=(self.dtype, 2),
+            )
+        else:
+            idx = int(self.end_id == col)  # 0 if col is start_id 1 otherwise.
+            self._data[:, idx] = np.fromiter(
+                (id_map[i] for i in self._data[:, idx]),
+                dtype=self.dtype,
+            )
+
     def _to_tsv(self, file_name, header):
         fmt = ["%d", "%d"]
         fmt.extend(["%f"] * len(self.features()))
@@ -119,7 +132,7 @@ class NumpyEdge(Edge):
         return self._data.copy()
 
     def as_igraph(self):
-        g = ig.Graph(self._data)
+        g = ig.Graph(self._data, directed=self.isdirected)
         for feat in self.features():
             g.es[feat] = self.feature_vector(feat)